fix meetup post permalinks
typical post permalinks are `/posts/meetups/YYYY/MM/DD/post-slug/`. the
problem is that we'll want to announce the meetup before we've got a
title. with the current setup, that would means that a meeting's url
would change from `/posts/meetups/YYYY/MM/DD/stay-tuned/` to
`/posts/meetups/YYYY/MM/DD/whatever-the-title-is/`. we'll omit the
`post-slug` part of the permalink so this doesn't change.