add coc to front page
going to go ahead and close the associated issue, because i don't have a
good plan for working location and organizers into the front page. plus
there's a pretty prominent link to the about page, which has the info,
so that's probably cool.

feel free to re-open if you've got ideas for these.

closes #13